Support version #s as strings, and remove some storages (#1355)
* JAX-RS generated classes for v2 of the Registry API.

* Update RuleViolationError.java

* Updated the API design.

* Version is now string everywhere.

* JAX-RS generated classes for v2 of the Registry API.

* Update RuleViolationError.java

* Started work on support for string versions

* work towards versions as strings

* Dropped the streams, asyncmem, and infinispan storages.  Updated remaining storages to support version-as-string

* removed the old in memory storages and various util classes that used to support the removed storages

* fix kubernetes tests

* minor test changes to help identify problems

* fix kubernetes tests manual trigger

* Replace old JBoss Inc copyright

* attempt to fix the logger test - only use java logging levels

* restore previous logger code

* added tests for versions as strings

* fix loggers test

* add missing -Pkafkasql in kubernetes tests

+++ app/src/main/java/io/apicurio/registry/storage/impl/SearchUtil.java
@@ -73,7 +73,7 @@ public class SearchUtil {
         searchedVersion.setName(artifactVersionMetaData.getName());
         searchedVersion.setState(artifactVersionMetaData.getState());
         searchedVersion.setType(artifactVersionMetaData.getType());
-        searchedVersion.setVersion((int) artifactVersionMetaData.getVersion());
+        searchedVersion.setVersion(artifactVersionMetaData.getVersion());
         searchedVersion.setLabels(artifactVersionMetaData.getLabels());
         return searchedVersion;
     }
